==> 0_parquet.ipynb <==
import pandas as pd
import numpy as np
%matplotlib inline
import joblib as jb

import json
import tqdm
import joblib
import gc

# simple parquet conversion to be fast
train = pd.read_json('./data/train_dataset.jl.gz',lines=True)
test = pd.read_json('./data/test_dataset.jl.gz',lines=True)
item_data = pd.read_json('./data/item_data.jl.gz',lines=True)

train.to_parquet("./data/train.parquet", engine='fastparquet', compression='GZIP')
test.to_parquet("./data/test.parquet", engine='fastparquet', compression='GZIP')
item_data.to_parquet("./data/item_data.parquet", engine='fastparquet', compression='GZIP')
